+++ tools/dpa-tools-ide-launchers/src/main/java/dk/statsbiblioteket/digital_pligtaflevering_aviser/tools/main/PDFContentLauncher.java
@@ -1,0 +1,26 @@
+package dk.statsbiblioteket.digital_pligtaflevering_aviser.tools.main;
+
+import dk.statsbiblioteket.digital_pligtaflevering_aviser.tools.maven.MavenProjectsHelper;
+
+import java.nio.file.Path;
+
+import static dk.statsbiblioteket.digital_pligtaflevering_aviser.tools.modules.BitRepositoryModule.BITREPOSITORY_SBPILLAR_MOUNTPOINT;
+import static dk.statsbiblioteket.medieplatform.autonomous.iterator.bitrepository.IngesterConfiguration.URL_TO_BATCH_DIR_PROPERTY;
+
+
+public class PDFContentLauncher {
+    public static void main(String[] args) throws Exception {
+        // for jaxb-impl 2.3.0 under Java 10 - https://github.com/javaee/jaxb-v2/issues/1197
+        System.setProperty("com.sun.xml.bind.v2.bytecode.ClassTailor.noOptimize", "true");
+
+
+        Path batchPath = MavenProjectsHelper.getRequiredPathTowardsRoot(PDFContentLauncher.class, "delivery-samples");
+        Path bitrepoPath = MavenProjectsHelper.getRequiredPathTowardsRoot(PDFContentLauncher.class, "bitrepositorystub-storage");
+
+        PDFContentMain.main(new String[]{
+                "pdf-content.properties",
+                BITREPOSITORY_SBPILLAR_MOUNTPOINT + "=" + bitrepoPath.toAbsolutePath(),
+                URL_TO_BATCH_DIR_PROPERTY + "=" + batchPath.toAbsolutePath()
+        });
+    }
+}