+++ tools/dpa-tools/src/test/java/dk/statsbiblioteket/digital_pligtaflevering_aviser/tools/main/ValidateXMLMainTest.java
@@ -42,7 +42,6 @@ public class ValidateXMLMainTest {
         xmlValidatorModule = new ValidateXMLMain.ValidateXMLModule();
     }
 
-
     // @org.junit.Test
     // FIXME:  This test assumes that all deliveries in source tree are _valid_.  Fails if adding more.  FIX by making an integration test.
     public void analyzeDeliveriesFolderTest() throws Exception {